requirements.txt
altgraph==0.17.2
beautifulsoup4==4.11.1
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.0.12
et-xmlfile==1.1.0
pandas==1.3.5
future==0.18.2
google-auth==2.8.0
google-auth-oauthlib==0.5.2
gspread==5.4.0
httplib2==0.20.4
idna==3.3
oauth2client==4.1.3
oauthlib==3.2.0
pefile==2022.5.30
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyinstaller-hooks-contrib==2022.7
pyparsing==3.0.9
pywin32-ctypes==0.2.0
requests==2.28.0
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
soupsieve==2.3.2.post1
urllib3==1.26.9
xmltodict==0.13.0
numpy==1.23.0
python-dotenv==0.21.0